Assignment_2/code/requirements.txt
confluent-kafka
cassandra-driver
argparse
pandas
flask